requirements.txt
asgiref==3.3.1
boto3==1.17.7
botocore==1.20.7
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.6
django-crispy-forms==1.11.0
django-heroku==0.3.1
django-storages==1.11.1
djangorestframework==3.12.2
docutils==0.16
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
mysqlclient==2.0.3
numpy==1.19.4
pandas==1.1.5
Pillow==8.1.0
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
users==1.0.dev0
whitenoise==5.2.0